Solution to Problem 1
After taking a good look at the code I found the part where the invoices are created, the code is similar for all the different billing schedules and the main difference is the number of invoices created according to billing_schedules.

Also, the bill dates for the monthly payments have an offset of only one month.