Problem 5 Solution
It looks like the problem creating a policy  with a billing schedule of 'Two-Pay' is related to a lack of definition of a 'Two-Pay' schedule in the make_invoices function. The logic to create a Two-Pay schedule is there but in billing_schedules is missing.

Since there is no logic to create a 'Semi-Annual' policy, I removed that from billing_schedules and changed it for a "Two-Pay" with only two invoices (maybe I'm wrong but I'm assuming that Two-Pay means that only two invoices are needed).

I also added some test cases for the Two-Pay schedule, I can see that some logic can be shared between the test cases so as an improvement I will probably add some helper functions to handle the repetitive code inside the tests cases.