Problem 6 Solution
This problem seems to be related to the a missing named_insured in the Policy 1, when this policy was created the named_insured was not set.

This may cause that when trying to make a payment, if the contact_id is not specified  the request to store the Payment data will fail since the contact_id field will be empty.

There are several possible ways to fix this, one of them is set a "default" contact_id but since I'm not sure about a valid default value I opted for the option of displaying an error message and return None, the caller should ensure that the function returns a valid Payment to verify that the payment was correct.

I also added some unit tests related to this functionality.