Problem 7 Solution
I based this solution on the one for the evaluate_cancel function.

I was not sure if this function should return True if the date_cursor was also higher than the cancel date, in my implementation this function only returns True if the date_cursor is higer than the due_date and lower than the cancel_date of the latest invoice.

I'm assuming that if an invoice passed the cancel date without payment, that policy will be canceled so it will be probably a good idea to call the evaluate_cancel function prior to evaluate_cancellation_pending_due_to_non_pay, since this last function won't be meaningful if the policy is already canceled.

I also added a small unit test for this change.