Problem 8 Solution
If I understand correctly I should not "delete" the invoices from the DB, just mark them as "deleted" in the corresponding field in the model Invoice and then create new invoices still based on the data of the policy.

In my change, the user will need to change the billing_schedule of the policy and then call the function make_invoices, this will flag the previous invoices as deleted and generate new ones with the correct amount for each one.

Another thing is that the old invoices are not accounted for anything like balance calculations, cancellation evaluations, etc.

A last thing that I noticed is that the invoice payment may have problems if the annual premium can't be divided exactly between the invoices, to solve this we can use floating numbers or charge an additional amount at the last payment of the policy to complete the annual premium.