Problem 7 Solution
If found some problems when doing some additional testing.

In the function evaluate_cancellation_pending_due_to_non_pay, if a new invoice is billed before or during the "cancellation" period, the code was going to return a True value even when the Invoice that it is in the cancellation range was already paid.