Problem 9
First part of the cancellation implementation, unit test still missing.

I based this implementation in the one for evaluate_cancellation_pending_due_to_non_pay, the original approach seems to fail when there are multiple invoices and the bill date of some of them is before the cancel date of the previous one.