Problem 10
I'm starting to work in the interface for the problem 10, I don't have a lot of experience in Knockout.js so I spent some time reading the documentation.

My plan is to add only a viewer at the moment, probably I will be able to add more functionality to create or edit the policies in future commits.