Problem 10
Added a basic implementation to view the Policies, this implementation only displays one policy at a time with the invoices for that policy, I think that this is te minimum that the problem requests.

I still need to add improvements to validate the data received by the server and how it should be handled (I guess that this is part of problem 11).

Finally, I think that it will be a good idea to add additional implementation to create or edit policies, I'm planning to add such implementation as part of the BONUS tasks.