BONUS
Just to add a small change for the day, I'm planning to include the option to edit data from the database directly on the web page.

At the moment, my plan is to add an option to create new policies and make payments to them, also evaluate cancellations.

But right now I'm just adding the button.