BONUS
Added more implementation to create new policies, still not making any changes to the database but I think that the data will be available to be used now.

I'm not really happy with this "monolithic" approach, I will probably need to separate the ViewModel in some way to make it more readable and modular.

The same applies for the 'index.html', maybe more templates are needed. This single page is getting to big.

Finally, I still need to improve the validations on the input data, specially now that I'm adding more functionality.