BONUS
Added functionality to create new policies.

1. It is possible to create a policy with new clients and agents, those will be added to the database after the policy creation.
2. The new policy will be added to the database and can be consulted using the existing functionality.

Notes:

I was thinking that probably is not a good idea to use a select for the policies and the clients, I imagine that those lists can grow a lot and make difficult to search for a specific value.

I still need to add the functionality to make payments and cancel policies.